requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
tensorflow-cpu>=2.10
Pillow>=9.0
python-dotenv>=1.0.0
python-dotenv>=1.0.0
openai